Add files via upload
Exploratory Data Analysis (EDA) on Titanic Dataset Titanic :
The Titanic dataset is a well-known dataset in the field of data science and machine learning. It contains information about passengers on board the Titanic, including whether they survived or not. The objective of this analysis is to explore the dataset, gain insights, and potentially discover patterns or trends.